+++ src/json_registry_manager/__init__.py
@@ -1,3 +1,3 @@
 """JSON Registry Manager."""
 
-__version__ = "0.1.2"
+__version__ = "0.1.3"